move getAgencies helper into a helper folder
